+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [CMS-1556] Fixed copy upstream not working for pages that have the same slug but exist under different apps
 - [CMS-1578] Fixed issue with `SSORedirectUsersToRequestAccessViews` middleware that caused an `AttributeError` to be raised when a users was not authenticated
+- [CMS-1423] Updated SSO notifications to use the user's first name in the greeting, instead of the full name
 - [CMS-1419] Fixed 'too_many_redirects' when logging in via SSO
 
 
@@ -13,6 +14,8 @@
 
 - [CMS-1419] Added Wagtail registration view for SSO users, group info management UI, and updated user edit view to mark users as approved
 - [CMS-1421] Integrate with GOV.UK Notify to send SSO user approval notifications
+- [CI-104] - Added capital invest landing, region, sector and opportunity pages.
+
 
 ## [2019.05.09](https://github.com/uktrade/directory-cms/releases/tag/2019.05.09)
 [Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.04.24_1...2019.05.09)